Move identity widget to separate script file
